EmpeddedId SuperiorTimeRecord

+++ tempera-server/src/main/java/at/qe/skeleton/rest/dtos/SuperiorTimeRecordDto.java
@@ -6,4 +6,4 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record SuperiorTimeRecordDto (UUID access_point_id, String tempera_station_id, LocalDateTime start, int duration, State mode, boolean auto_update){};
+public record SuperiorTimeRecordDto (UUID access_point_id, String tempera_station_id, LocalDateTime start, long duration, State mode, boolean auto_update){};
